requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
deepseek-ai==0.1.0  # 假设的 DeepSeek 包名，实际使用时请替换为正确的包名